procure_to_pay/utils.py
from rest_framework import viewsets, filters
from rest_framework.pagination import PageNumberPagination

class RequestPagination(PageNumberPagination):
    page_size = 10
    page_size_query_param = 'page_size'
    max_page_size = 100