requirements.txt
asgiref==3.11.0
attrs==25.4.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
PyPDF2==3.0.1
pypdfium2==5.0.0
pytesseract==0.3.10
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
wheel==0.45.1
reportlab==4.0.7
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
openai
google-genai>=0.1.0
celery==5.3.4
redis==5.0.1
